requirements.txt
pyyaml==6.0.1
gitpython==3.1.41
requests==2.31.0
packaging==24.0